add some size_t->uint casts in cases where we either know the length, or when it's unlikely to matter

+++ juno/base/events.d
@@ -167,7 +167,7 @@ struct Event(R, T...) {
   }
 
   private void addToList(TEventInfo e) {
-    uint n = list_.length;
+    size_t n = list_.length;
 
     if (n == 0)
       list_.length = 4;
